added roadway tests
